vmxws: Fix sending multiple-byte DT1 message
When a client requested address ranges by RQ1 or RQ0, the ranges are
saved. Usually, the host sends a DT1 message and it's address is
included in the saved range, the message will be forwarded to the
client.
However, when the start address of a DT1 message is outside of the
ranges, the message was ignored even if part of the DT1 address was
included in the range.